+++ package.json
@@ -18,6 +18,7 @@
     "cool-ascii-faces": "^1.3.4",
     "csv-read": "^2.0.2",
     "csvdata": "^2.0.0",
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "nedb": "^1.8.0"
   }
 }
